+++ src/main/java/appeng/parts/p2p/PartP2PRedstone.java
@@ -20,6 +20,8 @@ package appeng.parts.p2p;
 
 
 import appeng.api.parts.IPartModel;
+import appeng.api.parts.P2PTunnelModels;
+import appeng.core.AppEng;
 import appeng.items.parts.PartModels;
 import appeng.me.GridAccessException;
 import appeng.util.Platform;
@@ -29,6 +31,7 @@ import net.minecraft.block.state.IBlockState;
 import net.minecraft.item.ItemStack;
 import net.minecraft.nbt.NBTTagCompound;
 import net.minecraft.util.EnumFacing;
+import net.minecraft.util.ResourceLocation;
 import net.minecraft.util.math.BlockPos;
 import net.minecraft.world.IBlockAccess;
 import net.minecraft.world.World;
@@ -38,7 +41,7 @@ import java.util.List;
 
 public class PartP2PRedstone extends PartP2PTunnel<PartP2PRedstone> {
 
-    private static final P2PModels MODELS = new P2PModels("part/p2p/p2p_tunnel_redstone");
+    private static final P2PTunnelModels MODELS = new P2PTunnelModels(new ResourceLocation(AppEng.MOD_ID, "part/p2p/p2p_tunnel_redstone"));
 
     @PartModels
     public static List<IPartModel> getModels() {